+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 # Changelog
 
-All Notable changes to `laravel-notification-channels/webpus` will be documented in this file
+All Notable changes to `laravel-notification-channels/webpush` will be documented in this file
+
+## 1.0.0 - 2017-03-25
+
+- Added support for VAPID.
+- Added dedicated config file.
 
 ## 0.2.0 - 2017-01-26